integration: test 'MoveLeader' service
